requirements.txt
asgiref==3.3.0
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.3
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
